+++ modules/QingyuIpAgent/src/Providers/QingyuIpAgentServiceProvider.php
@@ -8,6 +8,10 @@ class QingyuIpAgentServiceProvider extends ServiceProvider
 {
     public function register(): void
     {
+        $this->mergeConfigFrom(
+            dirname(__DIR__, 2).DIRECTORY_SEPARATOR.'config'.DIRECTORY_SEPARATOR.'qingyu_ip_agent.php',
+            'qingyu_ip_agent'
+        );
     }
 
     public function boot(): void